Refactor tags:
- Move tag definitions into new `picard/const/tags.py` file.
- Move tag description generation into `TagVars` class.
- Update tests.
